requirements.txt
Flask
python-dotenv
azure-cognitiveservices-vision-computervision